+++ exponential.go
@@ -3,7 +3,7 @@ package backoff
 import (
 	"time"
 
-	"github.com/cenkalti/backoff"
+	"github.com/cenkalti/backoff/v4"
 )
 
 func NewExponential(maxWait, maxInterval time.Duration) BackOff {