+++ spec_test.go
@@ -3,7 +3,7 @@ package backoff
 import (
 	"testing"
 
-	"github.com/cenkalti/backoff"
+	"github.com/cenkalti/backoff/v4"
 )
 
 // Test_BackOff tests if this library and underlying implementation interfaces