+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Permanent backoff works with Golang error matching functions.
 
+### Changed
+- Bump github.com/cenkalti/backoff dependency from v2.2.1 to v4.0.2.
 
 ## [0.2.0] 2020-03-20